SMILE-419-upload-image-as-thumbnail, refactor to ChangeThumbnailButton and added translations
