+++ src/translations/translations_NO.json
@@ -226,6 +226,12 @@
     },
     "unpublished_resources": "Upubliserte ressurser"
   },
+  "thumbnail": {
+    "change_thumbnail": "Bytt miniatyrbilde",
+    "revert_thumbnail": "Slett opplastet og gå tilbake til foreslått miniatyrbilde",
+    "thumbnail_options": "Miniatyrbilde valg",
+    "upload_new_thumbnail": "Last opp nytt miniatyrbilde"
+  },
   "user": {
     "logged_in_as": "Logget på som"
   }
